+++ src/router.tsx
@@ -1,19 +1,19 @@
 import { Route, Routes } from 'react-router';
-import Components from './components';
-import SignUpLogin from './pages/SignUp/SignUpLogin';
-import SignUpEmail from './pages/SignUp/SignUpEmail';
-import SignUpNickName from './pages/SignUp/SignUpNickName';
-import SignUpInterests from './pages/SignUp/SignUpInterests';
+import Components from '@/components';
+import SignUp from '@/pages/SignUp/';
+import SignUpEmail from '@/pages/SignUp/Email';
+import SignUpNickname from '@/pages/SignUp/Nickname';
+import SignUpInterests from '@/pages/SignUp/Interests';
 import SignInPage from '@/pages/SignIn/Index';
 import ProfilePage from '@/pages/Profile/Index';
 import DiaryWrite from '@/pages/DiaryWrite/';
 
 export const navList = [
   { path: 'signin', text: '로그인', element: <SignInPage /> },
-  { path: 'signup_login', text: '회원가입(로그인 정보)', element: <SignUpLogin /> },
-  { path: 'signup_email', text: '회원가입(이메일)', element: <SignUpEmail /> },
-  { path: 'signup_nickname', text: '회원가입(닉네임)', element: <SignUpNickName /> },
-  { path: 'signup_interests', text: '회원가입(관심사)', element: <SignUpInterests /> },
+  { path: 'signup', text: '회원가입(로그인 정보)', element: <SignUp /> },
+  { path: 'signup/email', text: '회원가입(이메일)', element: <SignUpEmail /> },
+  { path: 'signup/nickname', text: '회원가입(닉네임)', element: <SignUpNickname /> },
+  { path: 'signup/interests', text: '회원가입(관심사)', element: <SignUpInterests /> },
   { path: 'diaryWrite', text: '일기 쓰기', element: <DiaryWrite /> },
   { path: 'profile', text: '프로필', element: <ProfilePage /> },
   { path: 'components', text: '컴포넌트 모음', element: <Components /> },
